pyproject.toml
[project]
name = "ilol-ladder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autopep8==2.0.2",
    "beautifulsoup4==4.11.2",
    "certifi==2022.12.7",
    "charset-normalizer==3.0.1",
    "idna==3.4",
    "pycodestyle==2.10.0",
    "python-dotenv==1.0.0",
    "requests==2.28.2",
    "riotwatcher>=3.3.1",
    "soupsieve==2.4",
    "tomli==2.0.1",
    "urllib3==1.26.14",
]
